prepared for merge into main
this commit un-commented a couple lines in the demo-scenario.yml and polished up the readme a bit. I'm going to try to bring a couple files (readme, demo-scenario-yml, custom-roles/add-user) from this commit into the main branch, but the goal is to leave my incomplete stuff (deploy-range.sh, rm-default-creds, etc) on this branch. We'll see what happens.